Fix pigiod name. Rename forecast-types to weather-types.
