+++ server/src/forecast-router.ts
@@ -1,8 +1,8 @@
 import { jsonOrJsonp } from './common';
 import { getForecast as getDsForecast } from './darksky-forecast';
 import { Request, Response, Router } from 'express';
-import { ForecastData } from './forecast-types';
 import { noCache } from './util';
+import { ForecastData } from './weather-types';
 import { getForecast as getWuForecast } from './wunderground-forecast';
 
 export const router = Router();
