+++ server/src/weather-types.ts
